+++ .travis.yml
@@ -34,8 +34,8 @@ script:
   - |
     php -l wp-cli-commands-stubs.php
   - |
-    php -l wp-cli-i18n-commands-stubs.php
+    php -l wp-cli-i18n-stubs.php
   - |
     git diff --exit-code
   - |
-    php -r 'require "vendor/php-stubs/wordpress-stubs/wordpress-stubs.php"; require "wp-cli-stubs.php";require "wp-cli-commands-stubs.php";require "wp-cli-i18n-commands-stubs.php";'
+    php -r 'require "vendor/php-stubs/wordpress-stubs/wordpress-stubs.php"; require "wp-cli-stubs.php";require "wp-cli-commands-stubs.php";require "wp-cli-i18n-stubs.php";'
