+++ generate.sh
@@ -4,13 +4,13 @@ HEADER=$'/**\n * Generated stub declarations for WP-CLI\n * @see https://wp-cli.
 
 FILE="wp-cli-stubs.php"
 FILE_PKGS="wp-cli-commands-stubs.php"
-FILE_PKGS_I18n="wp-cli-i18n-commands-stubs.php"
+FILE_I18N="wp-cli-i18n-stubs.php"
 
 set -e
 
 test -f "$FILE"
 test -f "$FILE_PKGS"
-test -f "$FILE_PKGS_I18n"
+test -f "$FILE_I18N"
 
 # Check wp-cli
 if [ ! -r ./source/vendor/ ]; then
@@ -50,10 +50,10 @@ fi
 # Packages.
 "$(dirname "$0")/vendor/bin/generate-stubs" \
     --force \
-    --finder=finder-commands-i18n.php \
+    --finder=finder-i18n.php \
     --header="$HEADER" \
     --functions \
     --classes \
     --interfaces \
     --traits \
-    --out="$FILE_PKGS_I18n"
+    --out="$FILE_I18N"
